# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: The Ministry of Herbology API
  version: 1.10.0
  description: |
    Contract for the Ministry of Herbology. Frozen — only the maintainers
    changes this file, via a recorded decision. Every path declares the parts of the project that
    implements it with `x-parts of the project`; `tests/contract/` enforces that the
    running API matches this document.

    Units: millimetres, °C, litres, lux. Timestamps are RFC 3339 UTC.
servers:
  - url: /api/v1
...
paths:
...
  /taxon/resolve:
    post:
      tags: [botany]
      summary: Resolve a typed name or photo to an accepted scientific name
      operationId: resolveTaxon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:      { type: string, description: Common or scientific name }
                image_key: { type: string, description: Uploaded photo for Pl@ntNet ID }
      responses:
        '200':
          description: Ranked candidate matches
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TaxonCandidate' }

  # ------------------------------------------------------------------ almanac
...
  /tending/rounds:
    get:
      tags: [tending]
      summary: Morning Rounds — what needs doing today
      operationId: getMorningRounds
      parameters:
        - { name: 'on', in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: Today's tasks, grouped
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MorningRounds' }
  /tending/tasks:
    get:
      tags: [tending]
      operationId: listTasks
      parameters:
        - { name: status,      in: query, schema: { $ref: '#/components/schemas/TaskStatus' } }
        - { name: specimen_id, in: query, schema: { type: string, format: uuid } }
        - { name: due_before,  in: query, schema: { type: string, format: date-time } }
        - { name: due_after,   in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
  /tending/tasks/{task_id}/complete:
    parameters:
      - { name: task_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [tending]
      summary: One-tap completion
      operationId: completeTask
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskCompletion' }
      responses:
        '200':
          description: Completed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /tending/tasks/complete-batch:
    post:
      tags: [tending]
      summary: Batch completion from Morning Rounds
      operationId: completeTasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task_ids]
              properties:
                task_ids:
                  type: array
                  items: { type: string, format: uuid }
                completed_by: { type: string, format: uuid }
      responses:
        '200':
          description: Completed tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
  /tending/care-rules:
    get:
      tags: [tending]
      operationId: listCareRules
      parameters:
        - { name: specimen_id, in: query, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Care rules
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CareRule' }
    post:
      tags: [tending]
      operationId: createCareRule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CareRule' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CareRule' }
  /tending/feeds:
    get:
      tags: [tending]
      summary: Calendar feeds for the household
      operationId: listCalendarFeeds
      responses:
        '200':
          description: Feeds
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CalendarFeed' }
    post:
      tags: [tending]
      summary: Create a tokenized ICS feed for a member and filter set
      operationId: createCalendarFeed
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CalendarFeedCreate' }
      responses:
        '201':
          description: Created; `webcal_url` is subscribable in Google and Apple Calendar
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CalendarFeed' }
  /tending/feeds/{feed_id}/revoke:
    parameters:
      - { name: feed_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [tending]
      operationId: revokeCalendarFeed
      responses:
        '204': { description: Revoked }
  /calendar/{token}.ics:
    parameters:
      - { name: token, in: path, required: true, schema: { type: string } }
    get:
      tags: [tending]
      summary: The ICS feed itself. Unauthenticated — the token is the secret.
      operationId: getCalendarIcs
      security: []
      responses:
        '200':
          description: iCalendar document
          content:
            text/calendar:
              schema: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }

  # ------------------------------------------------------------------ grounds
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: moh_session
    # the stack's own services, GET only. Nginx deletes this header
    # from every inbound request, so it is unusable from outside the stack.
    service:
      type: apiKey
      in: header
      name: X-Moh-Service-Token
    # the Lunette's device token, valid on /hub/lunette only.
    lunette:
      type: http
      scheme: bearer
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
        code:   { type: string }
    Confidence:
      type: string
      enum: [high, medium, low, unknown]
...
    Degradation:
      type: object
      description: >-
        One named reason an answer is worth less than a measurement. `caps_at`
        is a ceiling, not a subtraction: two doubts about the same input do not
        make it twice as doubtful, and neither may raise the other's ceiling.
        `detail` is a plain sentence, written to be shown to the reader as-is.
      required: [code, detail, caps_at]
      properties:
        code:    { type: string, description: 'Stable machine name, e.g. et0_fallback' }
        detail:  { type: string }
        caps_at: { $ref: '#/components/schemas/Confidence' }
    UnassessableSpecimen:
      type: object
      description: >-
        A plant whose risk could not be judged, and why. Served as
        `FrostReport.unassessable` and as `MorningRounds.unscheduled`.

        The panel that renders this exists to *name* the plants nobody could
        judge, so it carries the name. `specimen_id` alone forced every consumer
        to make a second request to `/specimens` just to avoid printing a uuid
        at a reader, and to print the uuid anyway when that request failed.
        the design.

        `specimen` is optional from 1.4.0 and **required in 2.0.0**, when
        `specimen_id` is retired.
        Populate it now; read it defensively until then.
      required: [specimen_id, reason]
      properties:
        specimen_id:
          type: string
          format: uuid
          deprecated: true
          description: 'Retired in 2.0.0 — read `specimen.id` instead.'
        specimen:    { $ref: '#/components/schemas/SpecimenBrief' }
        reason:      { type: string }
...
    TaskStatus:
      type: string
      enum: [due, done, satisfied, skipped, cancelled]
    TaskType:
      type: string
      enum: [water, fertilize, mist, prune, repot, inspect,
             bring_indoors, return_outdoors, cover, rotate]
...
    SpecimenBrief:
      type: object
      required: [id, display_name]
      properties:
        id:           { type: string, format: uuid }
        display_name: { type: string }
        is_outdoor:   { type: boolean }
        thumb_url:    { type: [string, 'null'] }
...
    SourceRef:
      type: object
      required: [id, kind, title]
      properties:
        id:      { type: string, format: uuid }
        kind:    { type: string }
        title:   { type: string }
        url:     { type: [string, 'null'] }
        license: { type: [string, 'null'] }
        retrieved_at: { type: string, format: date-time }
...
    TaxonCandidate:
      type: object
      required: [accepted_name, confidence]
      properties:
        accepted_name: { type: string }
        common_name:   { type: [string, 'null'] }
        # POWO and GBIF index species, not cultivars, so the cultivar is split
        # off the typed name during resolution and handed back here.
        cultivar:      { type: [string, 'null'] }
        family:        { type: [string, 'null'] }
        rank:          { type: string }
        gbif_key:      { type: [string, 'null'] }
        powo_id:       { type: [string, 'null'] }
        score:         { type: number }
        confidence:    { $ref: '#/components/schemas/Confidence' }
        source:        { $ref: '#/components/schemas/SourceRef' }
    ForecastPoint:
      type: object
      required: [time]
      properties:
        time:            { type: string, format: date-time }
        temperature_c:   { type: [number, 'null'] }
        temp_min_c:      { type: [number, 'null'] }
        temp_max_c:      { type: [number, 'null'] }
        precip_mm:       { type: [number, 'null'] }
        precip_prob_pct: { type: [number, 'null'] }
        et0_mm:          { type: [number, 'null'] }
        wind_kph:        { type: [number, 'null'] }
        condition:       { type: [string, 'null'] }
        sunrise:         { type: [string, 'null'], format: date-time }
        sunset:          { type: [string, 'null'], format: date-time }
...
    FrostAlert:
      type: object
      required:
        [id, specimen, night_of, forecast_low_c, threshold_c, action,
         confidence, degraded, degradations]
      properties:
        id:             { type: string, format: uuid }
        specimen:       { $ref: '#/components/schemas/SpecimenBrief' }
        night_of:       { type: string, format: date }
        forecast_low_c: { type: number }
        threshold_c:    { type: number }
        action:         { type: string, enum: [bring_indoors, cover, monitor] }
        advisory:       { type: [string, 'null'], description: 'NWS headline, if any' }
        task_id:
          type: [string, 'null']
          format: uuid
          description: >-
            The task raised for this alert, **populated only by the scheduler's
            surface** — `GET /tending/rounds`'s `alerts`. `GET /almanac/frost`
            serves `null` here and always will.

            That is a dependency direction, not an oversight. The alert is the
            weather engine's object and the task is the scheduler's; the
            scheduler reads the engine (`api/tending/` imports
            `almanac.service`), so the engine filling a field that points at a
            task would invert it. the design refuses that cycle for the
            irrigation term and it is refused here for the same reason: a cycle
            between the scheduler and the engine is how "why is this plant due?"
            stops having an answer anybody can trace.

            So a client that wants the alert *and* what is being done about it
            reads the round, which is the screen that question belongs to
            anyway. A client that wants only the risk reads the almanac.

            Whether this field should sit on `FrostAlert` at all is a 2.0.0
            question, with the rest of the breaking changes.
          x-populated-by: tending
        state:          { type: string, enum: [open, resolved, expired] }
        confidence:     { $ref: '#/components/schemas/Confidence' }
        degraded:       { type: boolean }
        degradations:
          type: array
          items: { $ref: '#/components/schemas/Degradation' }
...
    Task:
      type: object
      description: >-
        A task is the last link in a chain that starts at a care value and runs
        through the water balance. the design made a degraded `WaterBalance` say
        so; the design carries that the rest of the way, because the instruction
        a person acts on is where it matters most. `confidence`, `degraded` and
        `degradations` are required for the same reason they are on
        `WaterBalance`: an answer that may omit them cannot be told apart from
        one that had nothing to declare.
      required:
        [id, specimen, task_type, due_at, status, title, plain_title,
         confidence, degraded, degradations]
      properties:
        id:          { type: string, format: uuid }
        specimen:    { $ref: '#/components/schemas/SpecimenBrief' }
        task_type:   { $ref: '#/components/schemas/TaskType' }
        due_at:      { type: string, format: date-time }
        all_day:     { type: boolean }
        status:      { $ref: '#/components/schemas/TaskStatus' }
        satisfied_by: { type: [string, 'null'], enum: [rain, sensor, forecast_change, manual, null] }
        amount_ml:   { type: [integer, 'null'] }
        priority:    { type: string, enum: [low, normal, urgent] }
        title:       { type: string, description: 'Themed, e.g. "Tend the Monstera"' }
        plain_title: { type: string, description: 'Plain equivalent, e.g. "Water the Monstera — 500 ml"' }
        detail:      { type: [string, 'null'] }
        completed_at: { type: [string, 'null'], format: date-time }
        completed_by:
          # A task that is not done has no completer. Served as null by every
          # uncompleted task since 1.0.0, which a bare $ref to an object has
          # always rejected. the design.
          anyOf: [{ $ref: '#/components/schemas/Member' }, { type: 'null' }]
        deep_link:   { type: string, description: '/specimen/{id}/tending' }
        confidence:  { $ref: '#/components/schemas/Confidence' }
        degraded:    { type: boolean }
        degradations:
          type: array
          items: { $ref: '#/components/schemas/Degradation' }
          description: >-
            Each carries a plain sentence in `detail`, written to be shown to
            the reader as-is. A task built on an unattested coefficient must not
            arrive looking like a measurement.
      # the design — "satisfied" with no stated cause is the failure the design
      # was written about: an answer that reads as reassurance while carrying no
      # information. A screen should never have to say "the API did not say what
      # settled it". Checked here rather than asked for in prose.
      if:
        properties:
          status: { const: satisfied }
        required: [status]
      then:
        required: [satisfied_by]
        properties:
          satisfied_by:
            type: string
            enum: [rain, sensor, forecast_change, manual]
    TaskCompletion:
      type: object
      properties:
        completed_by: { type: string, format: uuid }
        amount_ml:    { type: integer }
        notes:        { type: string }
        new_location_id:
          type: string
          format: uuid
          description: Set when completing a bring_indoors task
    MorningRounds:
      type: object
      description: >-
        the design. A plant with no watering interval anywhere — no cited care
        value, no species column, no household rule — generates no task, and a
        plant with no task looks exactly like a plant that needs nothing.
        `unscheduled` is the same hole `/almanac/frost` filled with
        `unassessable`, in the screen people actually open every morning.
      required: [date, due, satisfied, alerts, unscheduled]
      properties:
        date:      { type: string, format: date }
        greeting:  { type: string }
        due:
          type: array
          items: { $ref: '#/components/schemas/Task' }
        satisfied:
          type: array
          items: { $ref: '#/components/schemas/Task' }
          description: Waterings the rain or a sensor already covered
        alerts:
          type: array
          items: { $ref: '#/components/schemas/FrostAlert' }
        unscheduled:
          type: array
          items: { $ref: '#/components/schemas/UnassessableSpecimen' }
          description: >-
            Plants the scheduler could not give a task, each with the reason.
            Silence has two meanings and this separates them.
        weather: { $ref: '#/components/schemas/ForecastPoint' }
    CareRule:
      type: object
      required: [task_type, strategy]
      properties:
        id:          { type: string, format: uuid }
        species_id:  { type: [string, 'null'], format: uuid }
        specimen_id: { type: [string, 'null'], format: uuid }
        task_type:   { $ref: '#/components/schemas/TaskType' }
        strategy:    { type: string, enum: [interval, water_balance, frost_guard, seasonal] }
        base_interval_days: { type: [integer, 'null'] }
        amount_ml:   { type: [integer, 'null'] }
        modifiers:   { type: object, additionalProperties: true }
        months:      { type: array, items: { type: integer } }
        enabled:     { type: boolean }
    CalendarFeed:
      type: object
      required: [id, name, webcal_url]
      properties:
        id:         { type: string, format: uuid }
        member_id:  { type: string, format: uuid }
        name:       { type: string }
        webcal_url: { type: string, description: 'webcal://host/api/v1/calendar/{token}.ics' }
        https_url:  { type: string }
        filters:    { $ref: '#/components/schemas/CalendarFilters' }
        push_target: { type: string, enum: [none, google, caldav] }
        push_status:
          type: string
          enum: ['off', awaiting_operator, ok, failing]
          description: >-
            Whether this feed's calendar push is working. `off`
            when `push_target` is `none`; `awaiting_operator` when it is
            `caldav` and the operator has not provisioned the
            `moh_caldav_<feed_id>` secret — a setup step, not a failure. 1.9.0.
        push_last_ok_at: { type: [string, 'null'], format: date-time }
        push_error:
          type: [string, 'null']
          description: >-
            The last push failure as a sentence an operator can act on. Never
            carries the CalDAV URL's credentials, the username or the password. 1.9.0.
        last_rendered_at: { type: [string, 'null'], format: date-time }
        revoked:    { type: boolean }
    CalendarFeedCreate:
      type: object
      required: [member_id, name]
      properties:
        member_id: { type: string, format: uuid }
        name:      { type: string }
        filters:   { $ref: '#/components/schemas/CalendarFilters' }
        push_target:
          type: string
          enum: [none, google, caldav]
          default: none
          description: >-
            `google` is refused with a 422 in 1.x: Google push needs an OAuth
            client and a per-member token store this project does not have. Subscribe to the ICS feed instead. `caldav` pushes
            once the operator provisions the feed's secret.
...
security:
  - session: []